// WADNR.Web/src/app/shared/generated/model/project-document-detail.ts
export class ProjectDocumentDetail { 
    ProjectDocumentID?: number;
    ProjectID?: number;
    DisplayName?: string | null;
    Description?: string | null;
    ProjectDocumentTypeID?: number | null;
    ProjectDocumentTypeDisplayName?: string | null;
    FileResourceID?: number;
    FileResourceGuid?: string | null;
    OriginalFileName?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}