// WADNR.Web/src/app/shared/generated/model/project-document-type-lookup-item.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectDocumentTypeLookupItem { 
    ProjectDocumentTypeID?: number;
    ProjectDocumentTypeDisplayName?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectDocumentTypeLookupItemForm { 
    ProjectDocumentTypeID?: FormControl<number>;
    ProjectDocumentTypeDisplayName?: FormControl<string>;
}

export class ProjectDocumentTypeLookupItemFormControls { 
    public static ProjectDocumentTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectDocumentTypeDisplayName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}