// WADNR.Web/src/app/shared/generated/model/project-document-upsert-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectDocumentUpsertRequest { 
    DisplayName: string;
    Description?: string | null;
    ProjectDocumentTypeID?: number | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectDocumentUpsertRequestForm { 
    DisplayName: FormControl<string>;
    Description?: FormControl<string>;
    ProjectDocumentTypeID?: FormControl<number>;
}

export class ProjectDocumentUpsertRequestFormControls { 
    public static DisplayName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
                Validators.minLength(1),
                Validators.maxLength(200),
            ],
        }
    );
    public static Description = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
                Validators.maxLength(1000),
            ],
        }
    );
    public static ProjectDocumentTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}